Fixed bugs

+++ lib/SpaceHandler/SpaceHandler.h
@@ -22,11 +22,11 @@
 
 
 struct Quaternion {
-    double w, x, y, z;
+    float w, x, y, z;
 };    
 
 struct Angles {
-    double from_x, from_y, from_z, theta;
+    float from_x, from_y, from_z, theta;
 };
 
 
@@ -38,7 +38,7 @@ Quaternion quaternion_normalize(Quaternion q);
 Quaternion quaternion_invert(Quaternion q);
 Quaternion quaternion_mult(Quaternion a, Quaternion b);        
 
-Angles get_angles_from_quat(Quaternion quat, uint8_t x_dir, uint8_t y_dir, uint8_t z_dir);
+Angles get_angles_from_quat(Quaternion quat, int8_t x_dir, int8_t y_dir, int8_t z_dir);
 
 
 #endif
